Entities split
