+++ src/st/nz0/e_axe_knight.c
