+++ src/st/nz0/e_bloody_skeleton.c
