+++ src/st/nz0/e_bloody_zombie.c
