+++ src/st/nz0/e_bone_scimitar.c
