+++ src/st/nz0/e_magically_sealed_door.c
