+++ src/st/nz0/e_skeleton.c
