+++ src/st/nz0/e_spittle_bone.c
